add: implementando segurança baseada no token do clerk

+++ src/routes/boards.js
@@ -2,6 +2,7 @@ const express = require('express');
 const router = express.Router();
 const controller = require('../controllers/boards');
 const { handleResponse } = require('../utils/output');
+const { clerkClient, getMembership, isUserTeam, getById } = require('../utils/clerk');
 
 /**
 * Create a new board for the specified team and board ID.
@@ -11,7 +12,13 @@ const { handleResponse } = require('../utils/output');
 * @param {Object} req.body - The request body containing board details, which will be stringified.
 * @returns {Promise<Object>} The result of the board creation operation.
 */
-router.post('/:team_id/:id', (req, res) => {
+router.post('/:team_id/:id', async (req, res) => {
+  // only member of team can do this
+  const { userId } = req.auth;
+  if (!(await isUserTeam(userId, teamId, 'org:member'))) {
+    return handleResponse(res, { status: 403, message: 'forbidden: only team members can perform this action', resource: `organization@${teamId}` });
+  }
+
   const result = controller.createBoard(req.params.team_id, req.params.id, JSON.stringify(req.body));
   handleResponse(res, result);
 });
@@ -24,7 +31,13 @@ router.post('/:team_id/:id', (req, res) => {
 * @param {Object} req.body - The request body containing board details, which will be stringified.
 * @returns {Promise<Object>} The result of the board creation operation.
 */
-router.get('/:team_id/:id', (req, res) => {
+router.get('/:team_id/:id', async (req, res) => {
+  // only member of team can do this
+  const { userId } = req.auth;
+  if (!(await isUserTeam(userId, teamId, 'org:member'))) {
+    return handleResponse(res, { status: 403, message: 'forbidden: only team members can perform this action', resource: `organization@${teamId}` });
+  }
+  
   const result = controller.getBoard(req.params.team_id, req.params.id);
   handleResponse(res, result);
 });
@@ -36,7 +49,13 @@ router.get('/:team_id/:id', (req, res) => {
 * @param {string} req.params.id - The ID of the board to delete.
 * @returns {Promise<Object>} The result of the board deletion operation.
 */
-router.delete('/:team_id/:id', (req, res) => {
+router.delete('/:team_id/:id', async (req, res) => {
+  // only admin of team can do this
+  const { userId } = req.auth;
+  if (!(await isUserTeam(userId, teamId))) {
+    return handleResponse(res, { status: 403, message: 'forbidden: only team admins can perform this action', resource: `organization@${teamId}` });
+  }
+
   const result = controller.deleteBoard(req.params.team_id, req.params.id);
   handleResponse(res, result);
 });
@@ -47,7 +66,13 @@ router.delete('/:team_id/:id', (req, res) => {
 * @param {string} req.params.team_id - The ID of the team for which to retrieve boards.
 * @returns {Promise<Object>} The result containing the list of boards for the specified team.
 */
-router.get('/:team_id', (req, res) => {
+router.get('/:team_id', async (req, res) => {
+  // only member of team can do this
+  const { userId } = req.auth;
+  if (!(await isUserTeam(userId, teamId, 'org:member'))) {
+    return handleResponse(res, { status: 403, message: 'forbidden: only team members can perform this action', resource: `organization@${teamId}` });
+  }
+  
   const result = controller.getTeamBoards(req.params.team_id);
   handleResponse(res, result);
 });
